+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### next version
 - preview's pattern is kept when changing file
+- "client-server" feature (see client-server.md)
 
 <a name="v0.19.1"></a>
 ### v0.19.1 - 2020-07-17